requirements.txt
opencv-python==4.6.0.66
numpy==1.26.4
Pillow==9.2.0